@@ -1,14 +1,14 @@
 name: flutter_country
-description: 'A lite flutter package for providing world Country Data without
-  internet connection, useful for country or language selector '
+description: "A lite flutter package for providing world Country Data without
+  internet connection, useful for country or language selector "
 version: 0.1.2
 homepage: https://github.com/egyleader/flutter_country.git
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  test: ^1.15.7
+  test: ^1.17.9
 
 dev_dependencies:
-  pedantic: ^1.11.0
+  pedantic: ^1.11.1
