@@ -1,13 +1,13 @@
-import 'package:flutter_country/models/timezone.dart';
+import 'timezone.dart';
 
 class TimeZoneList {
-  final List<TimeZone> timeZones;
+  final List<TimeZone>? timeZones;
 
   TimeZoneList({this.timeZones});
 
   factory TimeZoneList.fromJson(List<String> parsedJson) {
     var timeZones = <TimeZone>[];
-    if ( parsedJson == null ||  parsedJson.isEmpty) return TimeZoneList();
+    if (parsedJson.isEmpty) return TimeZoneList();
     parsedJson.forEach((timeZone) {
       timeZones.add(TimeZone.fromJson(timeZone));
     });
