@@ -1,125 +1,146 @@
-import 'package:flutter_country/models/country.dart';
-import 'package:flutter_country/assets/countries_data.dart';
-import 'package:flutter_country/models/country_list.dart';
-import 'package:flutter_country/models/timezone.dart';
+import 'assets/countries_data.dart';
 import 'models/country.dart';
 import 'models/country_list.dart';
+import 'models/timezone.dart';
 
 class Countries {
   // This class is not meant to be instantiated or extended; this constructor
   // prevents instantiation and extension.
-  Countries._();
-  static CountryList _countriesList;
 
-  static final CountryList _countryData = getCountries();
+  late CountryList _countriesList;
 
-  static CountryList getCountries() =>
-      _countriesList ?? CountryList.fromJson(countriesData);
-
-  static List<Country> all() {
-    return _countryData.countries;
+  Countries._() {
+    _countriesList = CountryList.fromJson(countriesData);
   }
 
-  static Country byName(String name) {
-    Country country = _countryData.countries.where((c) => c.name == name).first;
-    return country;
+  factory Countries() {
+    return Countries._();
   }
 
-  static Country byCode(String code) {
-    Country country =
-        _countryData.countries.where((c) => c.alpha2Code == code).first;
-    return country;
-  }
+  List<Country> get all => _countriesList.countries;
 
-  static Country byAlpha3Code(String code) {
-    Country country =
-        _countryData.countries.where((c) => c.alpha3Code == code).first;
-    return country;
-  }
-
-  static Country byNumericCode(String code) {
-    Country country =
-        _countryData.countries.where((c) => c.numericCode == code).first;
-    return country;
-  }
-
-  static Country byCallingCode(String code) {
-    Country country = _countryData.countries
-        .where((c) => c.callingCodes.contains(code))
+  Country byName(String name) {
+    Country country = _countriesList.countries
+        .where(
+          (c) => c.name == name,
+        )
         .first;
     return country;
   }
 
-  static Country byCapital(String capital) {
-    Country country =
-        _countryData.countries.where((c) => c.capital.contains(capital)).first;
+  Country byCode(String code) {
+    Country country = _countriesList.countries
+        .where(
+          (c) => c.alpha2Code == code,
+        )
+        .first;
     return country;
   }
 
-  static Country byFlag(String flag) {
-    Country country =
-        _countryData.countries.where((c) => c.flagIcon == flag).first;
+  Country byAlpha3Code(String code) {
+    Country country = _countriesList.countries
+        .where(
+          (c) => c.alpha3Code == code,
+        )
+        .first;
     return country;
   }
 
-  // static List<Country> byLanguageCode(String language) {
-  //   List<Country> countries = _countryData.countries
-  //       .where((c) => c.languages.containsKey(language))
-  //       .toList();
-  //   return countries;
-  // }
-
-  // static List<Country> byLanguageName(String language) {
-  //   List<Country> countries = _countryData.countries
-  //       .where((c) => c.languages.containsValue(language))
-  //       .toList();
-  //   return countries;
-  // }
-
-  static List<Country> unMembers() {
-    List<Country> countries =
-        _countryData.countries.where((c) => c.unMember == true).toList();
-    return countries;
+  Country byNumericCode(String code) {
+    Country country = _countriesList.countries
+        .where(
+          (c) => c.numericCode == code,
+        )
+        .first;
+    return country;
   }
 
-  static List<Country> independent() {
-    List<Country> countries =
-        _countryData.countries.where((c) => c.independent == true).toList();
-    return countries;
+  Country byCallingCode(String code) {
+    Country country = _countriesList.countries
+        .where(
+          (c) {
+            return c.callingCodes?.contains(code) ?? false;
+          },
+        )
+        .toList()
+        .first;
+    return country;
   }
 
-  static List<Country> byRegion(String region) {
-    List<Country> countries =
-        _countryData.countries.where((c) => c.region.contains(region)).toList();
-    return countries;
+  Country byCapital(String capital) {
+    Country country = _countriesList.countries.where(
+      (c) {
+        return c.capital?.contains(capital) ?? false;
+      },
+    ).first;
+    return country;
   }
 
-  static Country byArea(double area) {
-    Country countries =
-        _countryData.countries.where((c) => c.area == area).first;
-    return countries;
+  Country byFlag(String flag) {
+    Country country = _countriesList.countries
+        .where(
+          (c) => c.flagIcon == flag,
+        )
+        .first;
+    return country;
   }
 
-  static List<Country> areaBiggerThan(double area) {
-    List<Country> countries = _countryData.countries
-        .where((c) => c.area != null && c.area > area)
+  List<Country> unMembers() {
+    List<Country> countries = _countriesList.countries
+        .where(
+          (c) => c.unMember == true,
+        )
         .toList();
     return countries;
   }
 
-  static CountryList areaSmallerThan(double area) {
-    List<Country> countries = _countryData.countries
-        .where((c) => c.area != null && c.area < area)
+  List<Country> independent() {
+    List<Country> countries = _countriesList.countries
+        .where(
+          (c) => c.independent == true,
+        )
+        .toList();
+    return countries;
+  }
+
+  List<Country> byRegion(String region) {
+    List<Country> countries = _countriesList.countries
+        .where(
+          (c) => c.region?.contains(region) ?? false,
+        )
+        .toList();
+    return countries;
+  }
+
+  Country byArea(double area) {
+    Country countries =
+        _countriesList.countries.where((c) => c.area == area).first;
+    return countries;
+  }
+
+  List<Country> areaBiggerThan(double area) {
+    List<Country> countries = _countriesList.countries
+        .where((c) => c.area != null && c.area! > area)
+        .toList();
+    return countries;
+  }
+
+  CountryList areaSmallerThan(double area) {
+    List<Country> countries = _countriesList.countries
+        .where((c) => c.area != null && c.area! < area)
         .toList();
     return CountryList(countries: countries);
   }
 
-  static CountryList byTimeZone(TimeZone timeZone) {
-    List<Country> countries = _countryData.countries
-        .where((c) => c.timeZones.timeZones == null
+  CountryList byTimeZone(TimeZone timeZone) {
+    List<Country> countries = _countriesList.countries.where(
+      (c) {
+        if (c.timeZones == null) return false;
+        return c.timeZones!.timeZones == null
             ? false
-            : c.timeZones.timeZones.contains(timeZone))
-        .toList();
+            : c.timeZones!.timeZones!.contains(timeZone);
+      },
+    ).toList();
     return CountryList(countries: countries);
   }
 }
