@@ -1,5 +1,9 @@
 String getFlagByCountryCode(String code) {
-  String flag = code.toUpperCase().replaceAllMapped(RegExp(r'[A-Z]'),
-      (match) => String.fromCharCode(match.group(0).codeUnitAt(0) + 127397));
+  String flag = code.toUpperCase().replaceAllMapped(
+        RegExp(r'[A-Z]'),
+        (match) => String.fromCharCode(
+          match.group(0)!.codeUnitAt(0) + 127397,
+        ),
+      );
   return flag;
 }
